Implemented CRUD (clinet side) for employee form Added some validators Added snackbar for API response result Added state for save button Fixed some bugs

+++ frontend/src/app/employees/employees-list/employees-list.component.ts
@@ -56,12 +56,14 @@ export class EmployeesListComponent {
 
   openForm(id?: string) {
     if (!id) {
-      return this.router.navigate(['employee/new']);
+      return this.router.navigateByUrl('employee/new');
     }
-    return this.router.navigate([`employee/${id}/edit`]);
+    return this.router.navigateByUrl(`employee/${id}/edit`);
   }
 
   deleteEmployee(id: string): void {
-    this.employeeApi.delete(id).subscribe((employee) => { this.getAll() });
+    if (confirm('Are you sure you want to delete this employee?')) {
+      this.employeeApi.delete(id).subscribe((employee) => { this.getAll() });
+    }
   }
 }
